chore: refactor hash utilities
