+++ utils/hash/compare.js
