+++ utils/hash/file.js
