+++ utils/hash/index.js
@@ -1,0 +1,5 @@
+const file = require('./file');
+const files = require('./files');
+const compare = require('./compare');
+
+module.exports = { file, files, compare };
