+++ utils/index.js
@@ -2,9 +2,6 @@
 
 const sendCommand = require('./send-command');
 const logger = require('./logger');
-const calculateFileHash = require('./calculate-file-hash');
-const calculateFilesHash = require('./calculate-files-hash');
-const compareHashes = require('./compare-hashes');
 const fetchLatestVersion = require('./fetch-latest-version');
 const fetchPackageMeta = require('./fetch-package-meta');
 const incrementSemverVersion = require('./increment-semver-version');
@@ -15,9 +12,6 @@ const getCWD = require('./get-cwd');
 module.exports = {
     sendCommand,
     logger,
-    calculateFileHash,
-    calculateFilesHash,
-    compareHashes,
     fetchLatestVersion,
     fetchPackageMeta,
     incrementSemverVersion,
